stress-brk: remove redundant double cast on double variable

+++ stress-brk.c
@@ -345,10 +345,10 @@ static int stress_brk(stress_args_t *args)
 	pr_dbg("%s: %" PRIu64 " successful sbrk expands, %" PRIu64 " successful sbrk shinks\n",
 		args->name, brk_context->sbrk_expands, brk_context->sbrk_shrinks);
 
-	rate = (brk_context->sbrk_exp_count > 0.0) ? (double)brk_context->sbrk_exp_duration / brk_context->sbrk_exp_count : 0.0;
+	rate = (brk_context->sbrk_exp_count > 0.0) ? brk_context->sbrk_exp_duration / brk_context->sbrk_exp_count : 0.0;
 	stress_metrics_set(args, "nanosecs per sbrk page expand",
 		rate * STRESS_DBL_NANOSECOND, STRESS_METRIC_HARMONIC_MEAN);
-	rate = (brk_context->sbrk_shr_count > 0.0) ? (double)brk_context->sbrk_shr_duration / brk_context->sbrk_shr_count : 0.0;
+	rate = (brk_context->sbrk_shr_count > 0.0) ? brk_context->sbrk_shr_duration / brk_context->sbrk_shr_count : 0.0;
 	stress_metrics_set(args, "nanosecs per sbrk page shrink",
 		rate * STRESS_DBL_NANOSECOND, STRESS_METRIC_HARMONIC_MEAN);
